Rename favicon-yellow.png
favicon-yellow.png represents the canary favicon. This yellow
is almost totally indistinguishable from the normal orange icon for
colourblind GitLab users. This change renames it to
favicon-canary.png, which better represents what the favicon is used
for, and also changes it's appearance to make it easier for
colourblind users to distinguish.