Adds thread_safety style checks
In preparation for Puma, we need to start being more aware of thread
safety issues. This change adds the rubocop-thread_safety cop for
alerting developers and reviewers to potential thread-issues.

In this first version, many checks will be ignored, to be tackled in
future changes. Hopefully this will reduce the number of new thread
safety issues being introduced.