Rollback mutation of merge request on failed merge
The second call to the Gitaly UserMergeBranch RPC method can raise an
exception. Previously, if this were to happen, the merge request's
`in_progress_merge_commit_sha` attribute would be set to a commit SHA
that didn't exist.

`Repository#merge` now uses a transaction block in order to rollback
the setting of this attribute if any error occurs.

https://gitlab.com/gitlab-org/gitlab-ce/issues/60720