Non-blocking lock acquire for internal ids
This implements a non-blocking locking acquire strategy to update an internal id
record. It avoids statement timeouts when we previously were blocking
until we acquired the lock.

With this strategy, we use `FOR UPDATE NOWAIT` feature in PostgreSQL to
immediately return (unsuccessfully) from the lock acquire instead of
blocking on it. We sleep for 100ms and retry until we acquire the lock.

There is a drawback regarding locking fairness (for details see the code
comment)

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/61016