Attach the Puma Worker Killer through Lifecycle Events
This approach is the new favoured way of attaching lifecycle events, as
it does not require additional changes to be made to the puma.rb file
in GitLab-CE, Omnibus and Cloud Native GitLab. If these changes are not
synchronized, they may fail in certain environments.

Moving to lifecycle events avoids having to do this.

Additionally, this change will only install the puma-worker-killer
when Puma is running in multi-process mode. Puma can also run in single
process mode. In this mode, the puma-worker-killer would be ineffective.