Delay creating build's environment and deployment
Instead of creating environment and deployment on build creation add
this as a new prerequisite for builds. This way we are not going to
unnecessary create records (e.g. for builds with manual start).

Fix specs that assume environment and deployment will be created
immediately when build is created.

Related to https://gitlab.com/gitlab-org/gitlab-ce/issues/27644.