Fix feature tests
This makes sure the correct javascripts are loaded for the wiki pages
and wiki directories controllers.

It also makes sure that the wiki base path is passed to the markdown
service correctly, and includes tests to make sure that this does not
get out of sync.

It addresses issues with the performance of wiki pages on smaller page
sizes where the edit button was getting hidden.