Adds Multi-level cache
This change allows a multilevel cache to be configured in Rails.
When the GITLAB_L1_RAILS_CACHE_SIZE_MB is set, a local in-memory
cache, with the specified size, is configured in front of the Redis
cache.

This is done to reduce load on the redis cache, and also to improve
peformance.