New Cycle Analytics backend
This change lays the foundation for customizable cycle analytics stages.
The main reason for the change is to extract the event definitions to
separate objects (start_event, end_event) so that it could be easily
customized later on.

In this commit the default stages that we provide in CE are migrated to
the new structure.