Add experimental charts endpoint
That is not a working version of the code,
that's an endpoint prototype to test on localhost.

How to use:

```
http://localhost:3000/<PROJECT_NAMESPACE>/<PROJECT_ID>
/chart?url=<ENCODED_URL>
```

Where:
PROJECT_NAMESPACE and PROJECT_ID are valid project namespace,
and id on localhost. You can get them by calling
`Project.last` from Rails console.

ENCODED_URL is an Encoded URL to data provider.

To encode an URL you can use https://www.urlencoder.io

The endpoint responds with a JSON.
It's either `{ "error": "Error description" }`
or a valid JSON from data provider.